package/riscv-isa-sim: set --with-boost
Set --with-boost to avoid the following build failure when a non working
boost version is installed on host. This failure is raised since the
addition of the package in commit
d6aab7ae1607eb5c38ddb68e0cee997468b8c95d:

configure:5559: error: Could not find a version of the Boost::Asio library!

Fixes:
 - http://autobuild.buildroot.org/results/855b66b86d979cc922b947da932b0e2388481263

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>