support/scripts/gen-bootlin-toolchains: add support for new x86-64 toolchains
Following the merge of
99b69c0a3061aaa57a4508af51ae13f286d0f89d ("arch/Config.in.x86: add
option for -march=x86-64") and
5fcfefe4700913232e0c70c14f474f7039bee6d7 ("arch/Config.in.x86: add support for
x86-64-v2, x86-64-v3, x86-64-v4"), bootlin.toolchains.com now provides
toolchains targetting the x86-64, x86-64-v2, x86-64-v3 and x86-64-v4
architecture variants.

This commits modifies gen-bootlin-toolchains to support these
toolchains. It should be noted that the description for the x86-64-v3
and x86-64-v4 toolchains are for now the same, as Buildroot doesn't
yet have the options to describe the extra features that x86-64-v4
expects to find on the hardware platform.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>