package/boost: enable context/coroutine for arm variants
Add boost context support for aarch64/aarch64_be.

Context support for ARM is only available in the AAPCS ABI, so switch to
AAPCS ABI for all ARM variants.

While we're at it, update the URL about the context architecture support
to the one corresponding to the current version of boost.

Signed-off-by: Pieter De Gendt <pieter.degendt@gmail.com>
[Arnout: update URL, extend commit message]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>