package/ltp-testsuite: bump version to 20220121
LTP now builds metadata directory for the host. It supports
cross-compilation, we just need to pass HOST_CFLAGS and HOST_LDFLAGS in
addition to the target flags. They have to be passed to make, adding
them to configure does not work. Without this, on some architectures the
-mlongcalls and -mauto-litpools get passed to the host compile, but
they're not valid there.

Drop patches from this release.

Refresh 0001-lapi-Add-sysinfo.h-to-fix-build-with-MUSL-libc.patch
(still safe to carry it).

Signed-off-by: Petr Vorel <petr.vorel@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>