package/pkg-autotools: use /run as runtime path
Since many years now, /run is the standard runstate directory for Linux
systems, and Buildroot follows this standard. Encode this as the default
for autotools packages.

Note that the configure script bundled with a package may be generated
with an older version of Autotools, that may not yet support the
--runstatedir option. This will give an "Unsupported arguments" warning
and the option will be ignored. We may have better luck with passing
runstatedir as a variable instead of an option, but chances are that
the configure script doesn't know that variant either (it wouldn't print
a warning in that case). Since the warning is actually preferrable, pass
it as an option after all.

Signed-off-by: Norbert Lange <nolange79@gmail.com>
[Arnout: pass as an option instead of an argument after all]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>