package/systemd: remove hwdb-update service
Since the source HWDB files are already removed from the target, it
makes little sense to keep the service around.  If accidently
triggered it could create an empty database.

The service is installed in both staging and host directory as well and
not removed from there, so a user could craft his own solutions if it's
necessary to update the database.

Signed-off-by: Norbert Lange <nolange79@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>