configs/qemu_xtensa_lx60_nommu: use busybox minimal config
Update the qemu_xtensa_lx60_nommu_defconfig to use the
busybox-minimal.config.

After commit 42a35c08b9cc ("package/busybox: fix udhcpc options in minimal
config"), this has the benefit of fixing the following network
initialization failure:

  udhcpc: invalid option -- b

With the full busybox config, the -b option would still be passed and
udhcpc would fail to start for the reason above.

Note that on NOMMU, udhcpc backgrounds unconditionally (unless the -f
option is given), so it still behaves properly. The -b option in fact
only backgrounds after the lease is obtained; on NOMMU, backgrounding is
done before the lease is even requested. So the behaviour is more or
less the same, except that on MMU systems, networking can be considered
either up or not available after S20network, but on NOMMU there is no
such guarantee.

Signed-off-by: Vincent Stehlé <vincent.stehle@laposte.net>
Cc: Romain Naour <romain.naour@gmail.com>
Cc: Gerome Burlats <gerome.burlats@smile.fr>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>